Database Integration
There are issues currently with false errors being passed on signup and login pages even if all the information put in meets the requirements. In turn nothing is getting sent or retrieved from the database because an error is always created.

The creation of a database if one does not already exist i think has a better way of being done. such as hosting a database on a server and then referencing that so it isn't stored locally or always creating one if it can't find it. for the time being it will stay like this until the error message issue is resolved.